Prepping for Montreal town hall.
