+++ rajaDemo/src/main/java/com/kanawish/raja/BaselineRenderer.java
@@ -1,0 +1,34 @@
+package com.kanawish.raja;
+
+import android.content.Context;
+import android.support.annotation.Nullable;
+import android.view.MotionEvent;
+
+import org.rajawali3d.renderer.Renderer;
+
+import javax.microedition.khronos.egl.EGLConfig;
+import javax.microedition.khronos.opengles.GL10;
+
+abstract class BaselineRenderer extends Renderer {
+
+        final BaselineRenderingFragment exampleFragment;
+
+        public BaselineRenderer(Context context, @Nullable BaselineRenderingFragment fragment) {
+            super(context);
+            exampleFragment = fragment;
+        }
+
+        @Override
+        public void onRenderSurfaceCreated(EGLConfig config, GL10 gl, int width, int height) {
+            super.onRenderSurfaceCreated(config, gl, width, height);
+        }
+
+        @Override
+        public void onOffsetsChanged(float xOffset, float yOffset, float xOffsetStep, float yOffsetStep, int xPixelOffset, int yPixelOffset) {
+        }
+
+        @Override
+        public void onTouchEvent(MotionEvent event) {
+        }
+
+    }
