--- tangoRoomEditor/src/main/java/com/kanawish/raja/controlroom/PlanFittingApplication.java
@@ -1,15 +1,0 @@
-package com.kanawish.raja.controlroom;
-
-import android.app.Application;
-
-/**
- * Created by kanawish on 2016-07-17.
- */
-
-public class PlanFittingApplication extends Application {
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-    }
-}
